Separate Federation gherkin steps into trait
